+++ webapp/javascript/pages/tagExplorer/components/TotalSamplesChart/index.tsx
@@ -2,21 +2,28 @@ import React, { useMemo } from 'react';
 import { TimelineGroupData } from '@webapp/components/TimelineChart/TimelineChartWrapper';
 import { getFormatter } from '@pyroscope/flamegraph/src/format/format';
 import { Profile } from '@pyroscope/models/src';
+import LoadingSpinner from '@webapp/ui/LoadingSpinner';
 import PieChart, { PieChartDataItem } from './PieChart';
 import PieChartTooltip from './PieChartTooltip';
 import { calculateTotal } from '../../../math';
 import { formatValue } from '../../../formatTableData';
+import styles from './index.module.scss';
 
 interface TotalSamplesChartProps {
   filteredGroupsData: TimelineGroupData[];
   profile?: Profile;
   formatter?: ReturnType<typeof getFormatter>;
+  isLoading: boolean;
 }
+
+const CHART_HEIGT = '280px';
+const CHART_WIDTH = '280px';
 
 const TotalSamplesChart = ({
   filteredGroupsData,
   formatter,
   profile,
+  isLoading,
 }: TotalSamplesChartProps) => {
   const pieChartData: PieChartDataItem[] = useMemo(() => {
     return filteredGroupsData.length
@@ -28,12 +35,23 @@ const TotalSamplesChart = ({
       : [];
   }, [filteredGroupsData]);
 
+  if (!pieChartData.length || isLoading) {
+    return (
+      <div
+        style={{ width: CHART_WIDTH, height: CHART_HEIGT }}
+        className={styles.chartSkeleton}
+      >
+        <LoadingSpinner />
+      </div>
+    );
+  }
+
   return (
     <PieChart
       data={pieChartData}
       id="total-samples-chart"
-      height="280px"
-      width="280px"
+      height={CHART_HEIGT}
+      width={CHART_WIDTH}
       onHoverTooltip={(data) => (
         <PieChartTooltip
           label={data.label}