+++ webapp/javascript/services/apps.ts
@@ -1,4 +1,4 @@
-import { Apps, appsModel } from '@webapp/models/app';
+import { App, appsModel } from '@webapp/models/app';
 import { Result } from '@webapp/util/fp';
 import type { ZodError } from 'zod';
 import type { RequestError } from './base';
@@ -9,7 +9,7 @@ export interface FetchAppsError {
 }
 
 export async function fetchApps(): Promise<
-  Result<Apps, RequestError | ZodError>
+  Result<App[], RequestError | ZodError>
 > {
   const response = await request('/api/apps');
 
@@ -17,7 +17,7 @@ export async function fetchApps(): Promise<
     return parseResponse(response, appsModel);
   }
 
-  return Result.err<Apps, RequestError>(response.error);
+  return Result.err<App[], RequestError>(response.error);
 }
 
 export async function deleteApp(data: {