+++ packages/pyroscope-panel-plugin/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.13](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.12...@pyroscope/panel-plugin@1.4.13) (2022-11-30)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.4.12](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.11...@pyroscope/panel-plugin@1.4.12) (2022-11-29)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.4.11](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.10...@pyroscope/panel-plugin@1.4.11) (2022-11-28)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.4.10](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.9...@pyroscope/panel-plugin@1.4.10) (2022-11-28)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.4.9](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.8...@pyroscope/panel-plugin@1.4.9) (2022-11-23)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
 ## [1.4.8](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.7...@pyroscope/panel-plugin@1.4.8) (2022-11-21)
 
 **Note:** Version bump only for package @pyroscope/panel-plugin